add tools documentation to all tests and free variables
