+++ siliconcompiler/tools/bluespec/import.py
@@ -5,7 +5,8 @@ import shutil
 VLOG_DIR = 'verilog'
 
 def setup(chip):
-    ''' Sets up default settings on a per step basis
+    '''
+    Performs high level synthesis to generate a verilog output
     '''
 
     tool = 'bluespec'
