+++ siliconcompiler/tools/genfasm/genfasm.py
@@ -1,5 +1,9 @@
 '''
-To-Do: Add details here
+Generate a `FSAM <https://github.com/chipsalliance/fasm>`_ file from the output of `VPR <https://github.com/verilog-to-routing/vtr-verilog-to-routing>`_
+
+Documentation: https://docs.verilogtorouting.org/en/latest/utils/fasm/
+
+Sources: https://github.com/verilog-to-routing/vtr-verilog-to-routing/tree/master/utils/fasm
 '''
 
 ######################################################################
