+++ siliconcompiler/tools/magic/magic.py
@@ -9,6 +9,8 @@ Installation: https://github.com/RTimothyEdwards/magic
 Sources: https://github.com/RTimothyEdwards/magic
 '''
 
+import os
+
 ####################################################################
 # Make Docs
 ####################################################################
@@ -39,7 +41,7 @@ def setup(chip):
     chip.set('tool', tool, 'version', '>=8.3.196', clobber=False)
     chip.set('tool', tool, 'format', 'tcl')
 
-    chip.set('tool', tool, 'task', task, 'threads',  4, step=step, index=index, clobber=False)
+    chip.set('tool', tool, 'task', task, 'threads', os.cpu_count(), step=step, index=index, clobber=False)
     chip.set('tool', tool, 'task', task, 'refdir',  refdir, step=step, index=index, clobber=False)
     chip.set('tool', tool, 'task', task, 'script',  script, step=step, index=index, clobber=False)
