+++ siliconcompiler/tools/openroad/place.py
@@ -3,7 +3,8 @@ from siliconcompiler.tools.openroad.openroad import setup as setup_tool
 from siliconcompiler.tools.openroad.openroad import build_pex_corners, post_process
 
 def setup(chip):
-    ''' Helper method for configs specific to place tasks.
+    '''
+    Perform global and detail placements along with design violation repairs
     '''
 
     # Generic tool setup.
