+++ siliconcompiler/tools/sv2v/convert.py
@@ -1,7 +1,8 @@
+import os
 
 def setup(chip):
-    ''' Per tool function that returns a dynamic options string based on
-    the dictionary settings.
+    '''
+    Convert SystemVerilog to verilog
     '''
 
     chip.logger.debug("Setting up sv2v")
@@ -15,7 +16,7 @@ def setup(chip):
     chip.set('tool', tool, 'vswitch', '--numeric-version')
     chip.set('tool', tool, 'version', '>=0.0.9', clobber=False)
 
-    chip.set('tool', tool, 'task', task, 'threads',  4, step=step, index=index, clobber=False)
+    chip.set('tool', tool, 'task', task, 'threads', os.cpu_count(), step=step, index=index, clobber=False)
 
     # Since we run sv2v after the import/preprocess step, there should be no
     # need for specifying include dirs/defines. However we don't want to pass
