+++ siliconcompiler/tools/template/template.py
@@ -1,7 +1,7 @@
 '''
 Tool description
 
-Documentation:https://
+Documentation: https://
 
 Sources: https://
